refactor: #63 역할 기능 파일 분리
